+++ integration-test/playwright/page-objects/broker-protection.js
@@ -55,6 +55,7 @@ export class BrokerProtectionPage {
         await expect(this.page.getByLabel('First Name:')).toHaveValue('John')
         await expect(this.page.getByLabel('Last Name:')).toHaveValue('Smith')
         await expect(this.page.getByLabel('Phone Number:')).toHaveValue(/^\d{10}$/)
+        await expect(this.page.getByLabel('Street Address:')).toHaveValue(/^\d+ [A-Za-z]+(?: [A-Za-z]+)?$/)
         await expect(this.page.getByLabel('State:')).toHaveValue('IL')
         await expect(this.page.getByLabel('Zip Code:')).toHaveValue(/^\d{5}$/)
