+++ integration-test/test-pages/broker-protection/pages/form.html
@@ -27,6 +27,10 @@
         <input type="text" name="phone" id="user_phone">
     </label>
     <label>
+        Street Address:
+        <input type="text" name="street" id="user_street_address">
+    </label>
+    <label>
         State:
         <select name="state" id="state" class="cp-optout__form-input cp-optout__form-input--select">
             <option value="">All States</option>
