+++ src/features/broker-protection/actions/fill-form.js
@@ -1,5 +1,6 @@
 import { getElement, generateRandomInt } from '../utils.js'
 import { ErrorResponse, SuccessResponse } from '../types.js'
+import { generatePhoneNumber, generateZipCode, generateStreetAddress } from './generators.js'
 
 /**
  * @param {Record<string, any>} action
@@ -66,6 +67,8 @@ export function fillMany (root, elements, data) {
             }
 
             results.push(setValueForInput(inputElem, generateRandomInt(parseInt(element.min), parseInt(element.max)).toString()))
+        } else if (element.type === '$generated_street_address$') {
+            results.push(setValueForInput(inputElem, generateStreetAddress()))
         } else {
             if (!Object.prototype.hasOwnProperty.call(data, element.type)) {
                 results.push({ result: false, error: `element found with selector '${element.selector}', but data didn't contain the key '${element.type}'` })
@@ -176,23 +179,4 @@ function setImageUpload (element) {
         // failed
         return { result: false, error: e.toString() }
     }
-}
-
-export function generatePhoneNumber () {
-    /**
-     * 3 digits, 2-8, last two digits technically can't end in two 1s, but we'll ignore that requirement
-     * Source: https://math.stackexchange.com/questions/920972/how-many-different-phone-numbers-are-possible-within-an-area-code/1115411#1115411
-     */
-    const areaCode = generateRandomInt(200, 899).toString()
-
-    // 555-0100 through 555-0199 are for fictional use (https://en.wikipedia.org/wiki/555_(telephone_number)#Fictional_usage)
-    const exchangeCode = '555'
-    const lineNumber = generateRandomInt(100, 199).toString().padStart(4, '0')
-
-    return `${areaCode}${exchangeCode}${lineNumber}`
-}
-
-export function generateZipCode () {
-    const zipCode = generateRandomInt(10000, 99999).toString()
-    return zipCode
 }
